Load everything except categories on app start
